+++ crates/net/discv5/src/lib.rs
@@ -24,6 +24,7 @@ use tracing::{debug, error, trace};
 
 pub mod config;
 pub mod enr;
+pub mod error;
 pub mod filter;
 pub mod metrics;
 
@@ -31,43 +32,9 @@ pub use discv5::{self, IpMode};
 
 pub use config::{BootNode, Config, ConfigBuilder};
 pub use enr::uncompressed_id_from_enr_pk;
+pub use error::Error;
 pub use filter::{FilterOutcome, MustNotIncludeChains};
 use metrics::Discv5Metrics;
-
-/// Errors from using [`discv5::Discv5`] handle.
-#[derive(thiserror::Error, Debug)]
-pub enum Error {
-    /// Failure adding node to [`discv5::Discv5`].
-    #[error("failed adding node to discv5, {0}")]
-    AddNodeToDiscv5Failed(&'static str),
-    /// Missing key used to identify rlpx network.
-    #[error("fork missing on enr, 'eth' key missing")]
-    ForkMissing,
-    /// Failed to decode [`ForkId`] rlp value.
-    #[error("failed to decode fork id, 'eth': {0:?}")]
-    ForkIdDecodeError(#[from] alloy_rlp::Error),
-    /// Peer is unreachable over discovery.
-    #[error("discovery socket missing")]
-    UnreachableDiscovery,
-    /// Peer is unreachable over rlpx.
-    #[error("rlpx TCP socket missing")]
-    UnreachableRlpx,
-    /// Peer is not using same IP version as local node in discovery.
-    #[error("discovery socket is unsupported IP version, local ip mode: {0:?}")]
-    IpVersionMismatchDiscovery(IpMode),
-    /// Peer is not using same IP version as local node in rlpx.
-    #[error("rlpx TCP socket is unsupported IP version, local ip mode: {0:?}")]
-    IpVersionMismatchRlpx(IpMode),
-    /// Failed to initialize [`discv5::Discv5`].
-    #[error("init failed, {0}")]
-    InitFailure(&'static str),
-    /// An error from underlying [`discv5::Discv5`] node.
-    #[error("{0}")]
-    Discv5Error(discv5::Error),
-    /// An error from underlying [`discv5::Discv5`] node.
-    #[error("{0}")]
-    Discv5ErrorStr(&'static str),
-}
 
 /// Transparent wrapper around [`discv5::Discv5`].
 #[derive(Deref, DerefMut, Clone, Constructor)]
